rename(component): change Home2 to Home for clarity
