+++ src/App.jsx
@@ -1,10 +1,10 @@
-import Home2 from "./pages/Home2/Home2";
+import Home from "./pages/Home2/Home2";
 import { Routes, Route } from "react-router";
 function App() {
   return (
     <>
       <Routes>
-        <Route path="/" element={<Home2 />} key="home" />
+        <Route path="/" element={<Home />} key="home" />
       </Routes>
     </>
   );
