+++ src/pages/Home/Home.jsx
@@ -1,154 +1,114 @@
-import { GraduationCap, Globe2, Menu } from "lucide-react";
+import { Globe2, Users, MessageSquare, BookOpen, Trophy } from "lucide-react";
 import Plans from "../../component/Plans/Plans";
 import Email from "../../component/Email/Email";
-export default function Home() {
+import AboutSection from "../../component/Aboutsection/AboutSection";
+import ClientesLogos from "../../component/ClientsLogos/ClientesLogos";
+import StudentReviews from "../../component/StudentReviews/StudentReviews";
+import Footer from "../../component/Footer/Footer";
+import Header from "../../component/Header/Header";
+function Home2() {
   return (
-    <div className="min-h-screen bg-gradient-to-b from-blue-50 to-white">
-      <header className="fixed w-full bg-white/80 backdrop-blur-sm shadow-sm z-50">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex justify-between items-center h-16">
-            <div className="flex items-center">
-              <img src="./logo.png" width="50" height="50" alt="" />
-              <span className="ml-2 text-xl font-bold text-gray-900"></span>
-            </div>
-
-            <nav className="hidden md:flex space-x-8">
-              <a
-                href="#"
-                className="text-gray-700 hover:text-blue-600 px-3 py-2 text-sm font-medium"
-              >
-                Inicio
-              </a>
-              <a
-                href="/Home2"
-                className="text-gray-700 hover:text-blue-600 px-3 py-2 text-sm font-medium"
-              >
-                Home 2
-              </a>
-              <a
-                href="#"
-                className="text-gray-700 hover:text-blue-600 px-3 py-2 text-sm font-medium"
-              >
-                Sobre nosotros
-              </a>
-              <a
-                href="#Plans"
-                className="text-gray-700 hover:text-blue-600 px-3 py-2 text-sm font-medium"
-              >
-                Nuestros Planes
-              </a>
-              <a
-                href="#"
-                className="text-gray-700 hover:text-blue-600 px-3 py-2 text-sm font-medium"
-              >
-                Contacto
-              </a>
-            </nav>
-
-            <div className="md:hidden">
-              <button className="p-2 rounded-md text-gray-700 hover:text-blue-600">
-                <Menu className="h-6 w-6" />
-              </button>
-            </div>
-          </div>
-        </div>
-      </header>
-
-      <div className="relative pt-16">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 pt-20 pb-16 lg:pt-12">
-          <div className="grid lg:grid-cols-2 gap-12 items-center">
-            <div>
-              <div className="inline-flex items-center px-4 py-2 rounded-full bg-blue-100 text-blue-700 mb-6">
-                <Globe2 className="h-5 w-5 mr-2" />
-                <span className="text-sm font-medium">
-                  Clases presenciales y online
-                </span>
-              </div>
-
-              <h1 className="text-4xl lg:text-5xl font-bold text-gray-900 leading-tight mb-6">
-                Domina el inglés y abre un mundo de oportunidades 🚀
-              </h1>
-
-              <p className="text-xl text-gray-600 mb-8 leading-relaxed">
-                Aprende de manera rápida, divertida y efectiva con nuestro
-                método probado. ¡Comienza hoy y transforma tu futuro con
-                profesores nativos certificados!
-              </p>
-
-              <div className="flex flex-col sm:flex-row gap-4">
-                <button className="px-8 py-4 bg-blue-600 text-white rounded-lg font-semibold hover:bg-blue-700 transition-colors">
-                  Comienza ahora
-                </button>
-                <button className="px-8 py-4 bg-white text-blue-600 rounded-lg font-semibold border-2 border-blue-600 hover:bg-blue-50 transition-colors">
-                  Clase de prueba gratis
-                </button>
-              </div>
-
-              <div className="mt-8 flex items-center gap-6">
-                <div className="flex -space-x-2">
-                  <img
-                    className="w-10 h-10 rounded-full border-2 border-white"
-                    src="https://images.unsplash.com/photo-1494790108377-be9c29b29330?auto=format&fit=crop&w=100&h=100"
-                    alt="Student"
-                  />
-                  <img
-                    className="w-10 h-10 rounded-full border-2 border-white"
-                    src="https://images.unsplash.com/photo-1507003211169-0a1dd7228f2d?auto=format&fit=crop&w=100&h=100"
-                    alt="Student"
-                  />
-                  <img
-                    className="w-10 h-10 rounded-full border-2 border-white"
-                    src="https://images.unsplash.com/photo-1438761681033-6461ffad8d80?auto=format&fit=crop&w=100&h=100"
-                    alt="Student"
-                  />
-                  <img
-                    className="w-10 h-10 rounded-full border-2 border-white"
-                    src="https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?auto=format&fit=crop&w=100&h=100"
-                    alt="Student"
-                  />
+    <div className="min-h-screen bg-white">
+      <Header />
+      <main className="pt-16">
+        <section className="py-20 bg-gradient-to-b from-indigo-50 to-white">
+          <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+            <div className="grid lg:grid-cols-2 gap-12 items-center">
+              <div>
+                <div className="inline-flex items-center px-4 py-2 rounded-full bg-indigo-100 text-indigo-700 mb-6">
+                  <Globe2 className="h-5 w-5 mr-2" />
+                  <span className="text-sm font-medium">
+                    Aprende inglés de forma natural
+                  </span>
                 </div>
-                <div>
-                  <div className="text-blue-600 font-bold">
-                    +1000 estudiantes
+
+                <h1 className="text-4xl lg:text-5xl font-bold text-gray-900 leading-tight mb-6">
+                  Domina el inglés con profesores nativos certificados 🎓
+                </h1>
+
+                <p className="text-xl text-gray-600 mb-8">
+                  Nuestro método único te ayuda a pensar y hablar en inglés de
+                  forma natural. Aprenderás el idioma como lo hacen los nativos,
+                  con ejercicios prácticos y conversación real.
+                </p>
+
+                <div className="flex flex-col sm:flex-row gap-4 mb-12">
+                  <button className="px-8 py-4 bg-indigo-600 text-white rounded-lg font-semibold hover:bg-indigo-700 transition-colors">
+                    Empieza tu viaje
+                  </button>
+                  <button className="px-8 py-4 bg-white text-indigo-600 rounded-lg font-semibold border-2 border-indigo-600 hover:bg-indigo-50 transition-colors">
+                    Clase de prueba gratis
+                  </button>
+                </div>
+
+                <div className="flex items-center gap-8 text-gray-600">
+                  <div className="flex items-center gap-2">
+                    <Users className="h-5 w-5 text-indigo-600" />
+                    <span>+5000 estudiantes</span>
                   </div>
-                  <div className="text-gray-600 text-sm">
-                    confían en nosotros
+                  <div className="flex items-center gap-2">
+                    <Trophy className="h-5 w-5 text-indigo-600" />
+                    <span>98% tasa de éxito</span>
                   </div>
                 </div>
               </div>
-            </div>
 
-            <div className="relative">
-              <div className="relative rounded-2xl overflow-hidden shadow-2xl">
-                <img
-                  src="/photo.avif"
-                  alt="Students learning"
-                  className="w-full h-[600px] object-cover"
-                />
-                <div className="absolute inset-0 bg-gradient-to-t from-black/20 to-transparent"></div>
-              </div>
+              <div className="relative">
+                <div className="relative rounded-2xl overflow-hidden bg-white p-8 shadow-xl">
+                  <img
+                    src="https://images.unsplash.com/photo-1522202176988-66273c2fd55f?ixlib=rb-4.0.3&auto=format&fit=crop&w=1000&q=80"
+                    alt="Students learning"
+                    className="w-full h-[400px] object-cover rounded-lg"
+                  />
+                </div>
 
-              <div className="absolute -bottom-6 -left-6 bg-white rounded-lg shadow-xl p-4 w-64">
-                <div className="flex items-center gap-3 mb-2">
-                  <div className="bg-yellow-100 p-2 rounded-lg">
-                    <GraduationCap className="h-6 w-6 text-yellow-600" />
-                  </div>
-                  <div>
-                    <div className="font-semibold text-gray-900">
-                      98% de éxito
+                <div className="absolute -left-8 top-1/2 bg-white rounded-lg shadow-xl p-4 w-64">
+                  <div className="flex items-center gap-3">
+                    <div className="bg-green-100 p-2 rounded-lg">
+                      <MessageSquare className="h-6 w-6 text-green-600" />
                     </div>
-                    <div className="text-sm text-gray-600">
-                      en certificaciones
+                    <div>
+                      <div className="font-semibold text-gray-900">
+                        Conversación Real
+                      </div>
+                      <div className="text-sm text-gray-600">
+                        Práctica con nativos
+                      </div>
+                    </div>
+                  </div>
+                </div>
+
+                <div className="absolute -right-8 bottom-12 bg-white rounded-lg shadow-xl p-4 w-64">
+                  <div className="flex items-center gap-3">
+                    <div className="bg-blue-100 p-2 rounded-lg">
+                      <BookOpen className="h-6 w-6 text-blue-600" />
+                    </div>
+                    <div>
+                      <div className="font-semibold text-gray-900">
+                        Método Efectivo
+                      </div>
+                      <div className="text-sm text-gray-600">
+                        Aprende más rápido
+                      </div>
                     </div>
                   </div>
                 </div>
               </div>
             </div>
           </div>
-        </div>
-      </div>
-      <Plans />
-      <Email />
+        </section>
+
+        <ClientesLogos />
+
+        <AboutSection />
+
+        <Plans />
+        <StudentReviews />
+        <Email />
+      </main>
+      <Footer />
     </div>
   );
 }
+
+export default Home2;
